+++ src/module.database/provider.memory/memory.database.spec.ts
@@ -1,7 +1,7 @@
 import { Test } from '@nestjs/testing'
 import { MemoryDatabaseModule } from '@src/module.database/provider.memory/index'
 import { MemoryDatabase } from '@src/module.database/provider.memory/memory.database'
-import * as spec from '@src/module.database/database.spec/database.spec'
+import * as spec from '@src/module.database/database.spec/specifications'
 
 let database: MemoryDatabase
